# docs/swagger.yaml
basePath: /api
definitions:
...
  models.CreateTransactionRequest:
    properties:
      account_id:
        type: integer
      amount:
        type: number
    type: object
...
  models.GenerateSOACustomRequest:
    properties:
      account_id:
        type: integer
      currency:
        default: USD
        type: string
      end_date:
        type: string
      item_count:
        default: 100
        type: integer
      start_date:
        type: string
    type: object
...
  models.Response:
    properties:
      data: {}
      message:
        type: string
      status_code:
        type: integer
      success:
        type: boolean
    type: object
  models.SOA:
    properties:
      account_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      pdf_url:
        type: string
      statement_date:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a Banking System API.
  title: Banking System API
  version: "1.0"
paths:
...
  /soa/download:
    get:
      consumes:
      - application/json
      description: Download a statement of account by ID
      parameters:
      - description: SOA ID
        in: query
        name: id
        required: true
        type: integer
      - default: Bearer <Add access token here>
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: SOA downloaded successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Download Statement of Account
      tags:
      - soa
  /soa/generate:
    post:
      consumes:
      - application/json
      description: Get a statement of account for a user with custom filters such
        as start date, end date, transaction type and item count.
      parameters:
      - description: Generate SOA Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.GenerateSOACustomRequest'
      - default: Bearer <Add access token here>
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SOA'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Get Statement of Account
      tags:
      - soa
  /soa/generated:
    get:
      consumes:
      - application/json
      description: Get all statements of account generated by the user
      parameters:
      - default: Bearer <Add access token here>
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Get All Generated SOAs
      tags:
      - soa
  /transaction:
    get:
      consumes:
      - application/json
      description: Get all transactions with optional filtering, sorting, and pagination
      parameters:
      - in: query
        name: account_id
        type: integer
      - in: query
        name: date_from
        type: string
      - in: query
        name: date_to
        type: string
      - in: query
        name: max_amount
        type: number
      - in: query
        name: min_amount
        type: number
      - enum:
        - pending
        - completed
        - failed
        in: query
        name: status
        type: string
        x-enum-varnames:
        - Pending
        - Completed
        - Failed
      - enum:
        - DEPOSIT
        - WITHDRAWAL
        in: query
        name: type
        type: string
        x-enum-varnames:
        - Deposit
        - Withdrawal
      - description: '"ASC" or "DESC"'
        in: query
        name: direction
        type: string
      - in: query
        name: field
        type: string
      - in: query
        name: page
        type: integer
      - in: query
        name: page_size
        type: integer
      - default: Bearer <Add access token here>
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Get all transactions of authenticated user
      tags:
      - transactions
  /transaction/deposit:
    post:
      consumes:
      - application/json
      description: Make a deposit transaction
      parameters:
      - description: Deposit details
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/models.CreateTransactionRequest'
      - default: Bearer <Add access token here>
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Deposit money into an account
      tags:
      - transactions
  /transaction/get:
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific transaction
      parameters:
      - description: Transaction ID
        in: query
        name: id
        required: true
        type: integer
      - default: Bearer <Add access token here>
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Get a specific transaction of authenticated user
      tags:
      - transactions
  /transaction/withdraw:
    post:
      consumes:
      - application/json
      description: Make a withdrawal transaction
      parameters:
      - description: Withdrawal details
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/models.CreateTransactionRequest'
      - default: Bearer <Add access token here>
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Withdraw money from an account
      tags:
      - transactions
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"